Extracted re-usable logic to produce test migration records
